requirements.txt
streamlit
streamlit-chat
python-dotenv
langchain
openai
Qdrant
pinecone-client